+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.48 ] - 2024/06/08
+
+  * Fixed a bug that was causing `ValueError: list.remove(x): x not in list` exceptions to be raised whenever the user changed configuration options for a device.
+
 ###### [ 1.0.47 ] - 2024/06/07
 
   * Updated underlying `spotifywebapiPython` package requirement to version 1.0.46.